Move transaction-related types into `transaction` module
