+++ crates/contracts/core/ab-contracts-common/src/lib.rs
@@ -7,6 +7,7 @@ pub mod env;
 mod error;
 pub mod metadata;
 pub mod method;
+pub mod transaction;
 
 use crate::method::MethodFingerprint;
 use ab_contracts_io_type::IoType;
