+++ crates/contracts/core/ab-contracts-test-utils/src/transaction_builder.rs
@@ -1,6 +1,7 @@
 use ab_contracts_common::Address;
-use ab_contracts_common::env::{Blake3Hash, Gas, Transaction, TransactionHeader, TransactionSlot};
+use ab_contracts_common::env::Blake3Hash;
 use ab_contracts_common::method::ExternalArgs;
+use ab_contracts_common::transaction::{Gas, Transaction, TransactionHeader, TransactionSlot};
 use ab_system_contract_simple_wallet_base::payload::TransactionMethodContext;
 use ab_system_contract_simple_wallet_base::payload::builder::{
     TransactionPayloadBuilder, TransactionPayloadBuilderError,
