+++ crates/contracts/system/ab-system-contract-simple-wallet-base/src/seal.rs
@@ -2,7 +2,8 @@
 
 use crate::{SIGNING_CONTEXT, Seal};
 use ab_contracts_common::ContractError;
-use ab_contracts_common::env::{Blake3Hash, TransactionHeader, TransactionSlot};
+use ab_contracts_common::env::Blake3Hash;
+use ab_contracts_common::transaction::{TransactionHeader, TransactionSlot};
 use ab_contracts_io_type::trivial_type::TrivialType;
 use core::slice;
 use schnorrkel::context::SigningContext;
